refactor: move example out of test
