+++ packages/blade/src/components/SideNav/__tests__/SideNav.web.test.tsx
@@ -1,97 +1,8 @@
 import userEvents from '@testing-library/user-event';
-import { Link, MemoryRouter, useLocation, matchPath } from 'react-router-dom';
 import { waitFor } from '@testing-library/react';
-import {
-  SideNav,
-  SideNavBody,
-  SideNavFooter,
-  SideNavSection,
-  SideNavLink,
-  SideNavItem,
-  SideNavLevel,
-} from '../index';
-import type { SideNavLinkProps, SideNavProps } from '../index';
-import { Indicator } from '~components/Indicator';
-import { Switch } from '~components/Switch';
-import renderWithTheme from '~utils/testing/renderWithTheme.web';
-import { HomeIcon, PaymentGatewayIcon, PaymentLinkIcon } from '~components/Icons';
-
-const isItemActive = (
-  location: { pathname: string },
-  { href, activeOnLinks }: { href?: string; activeOnLinks?: string[] },
-): boolean => {
-  const isCurrentPathActive = Boolean(
-    matchPath(location.pathname, {
-      path: href,
-      exact: false,
-    }),
-  );
-
-  const isSubItemActive = Boolean(
-    activeOnLinks?.find((href) => matchPath(location.pathname, { path: href, exact: false })),
-  );
-
-  return isCurrentPathActive || isSubItemActive;
-};
-
-const NavLink = (
-  props: Omit<SideNavLinkProps, 'as'> & {
-    activeOnLinks?: string[];
-  },
-): React.ReactElement => {
-  const location = useLocation();
-
-  return (
-    <SideNavLink
-      {...props}
-      as={Link}
-      isActive={isItemActive(location, { href: props.href, activeOnLinks: props.activeOnLinks })}
-    />
-  );
-};
-
-const SideNavExample = ({
-  initialEntries = ['/home'],
-  ...args
-}: Omit<SideNavProps, 'children'> & { initialEntries?: string[] }): React.ReactElement => {
-  return (
-    <MemoryRouter initialEntries={initialEntries}>
-      <SideNav {...args}>
-        <SideNavBody>
-          <NavLink icon={HomeIcon} title="Home" href="/home" />
-          <NavLink
-            title="L2 Trigger"
-            href="/l2-item"
-            activeOnLinks={['/l2-item', '/l2-item-2', '/l3-item', '/l3-item-2']}
-          >
-            <SideNavLevel>
-              <NavLink title="L2 Item" href="/l2-item" />
-              <NavLink title="L2 Item 2" href="/l2-item-2" />
-              <NavLink title="L3 Trigger" activeOnLinks={['/l3-item', '/l3-item-2']}>
-                <SideNavLevel>
-                  <NavLink title="L3 Item" href="/l3-item" />
-                  <NavLink title="L3 Item 2" href="/l3-item-2" />
-                </SideNavLevel>
-              </NavLink>
-            </SideNavLevel>
-          </NavLink>
-
-          <SideNavSection title="Products">
-            <NavLink icon={PaymentGatewayIcon} title="Gateway" href="/gateway" />
-            <NavLink icon={PaymentLinkIcon} title="Links" href="/links" />
-          </SideNavSection>
-        </SideNavBody>
-        <SideNavFooter>
-          <SideNavItem
-            leading={<Indicator color="positive" accessibilityLabel="" />}
-            title="Test Mode"
-            trailing={<Switch accessibilityLabel="" />}
-          />
-        </SideNavFooter>
-      </SideNav>
-    </MemoryRouter>
-  );
-};
+import { SideNavExample } from './SideNavExample';
+import renderWithTheme from '~utils/testing/renderWithTheme';
+import assertAccessible from '~utils/testing/assertAccessible';
 
 describe('SideNav', () => {
   test('should render', () => {
@@ -109,9 +20,7 @@ describe('SideNav', () => {
 
   test('should open L2 and L3', async () => {
     const user = userEvents.setup();
-    const { getByRole } = renderWithTheme(<SideNavExample display="block" />, {
-      container: document.body,
-    });
+    const { getByRole } = renderWithTheme(<SideNavExample display="block" />);
 
     // Open L2
     await user.click(getByRole('link', { name: 'L2 Trigger' }));
@@ -134,10 +43,14 @@ describe('SideNav', () => {
   test('should keep L3 Item selected based on URL', () => {
     const { getByRole } = renderWithTheme(
       <SideNavExample display="block" initialEntries={['/l3-item']} />,
-      {
-        container: document.body,
-      },
     );
+
     expect(getByRole('link', { name: 'L3 Item' })).toHaveAttribute('aria-current', 'page');
+  });
+
+  test('should pass general a11y', async () => {
+    const { container } = renderWithTheme(<SideNavExample display="block" />);
+
+    await assertAccessible(container);
   });
 });
